refactor: :lock: Finalizado backend  microsserviço catálogo de vídeos
Foi fnalizado todas as alterações no microsserviço de catálogo de vídeos, com relação ao backend

+++ .docker/entrypoint.sh
@@ -1,9 +1,19 @@
 #!/bin/bash
 
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
-# chown -R www-data:www-data .
-cp .env.example .env
-cp .env.testing.example .env.testing
+
+###FRONT-END
+npm config set cache /var/www/.npm-cache --global
+cd /var/www/frontend && npm install && cd ..
+
+###BACK-END
+cd backend
+if [ ! -f ".env"]; then
+    cp .env.example .env
+fi
+if [ ! -f ".env.testing"]; then
+    cp .env.testing.example .env.testing
+fi
 composer install
 # composer install --no-interaction --prefer-dist --optimize-autoloader -d /var/www
 php artisan config:cache
